fix: rename issue-links.tsx
